clean up obs (#50)
We previously defined an inline observation function per object type, and then needed to use _type_id to figure out what casting to do in observation_encoder. I believe we had to do that since we'd been writing all functions in cython header files, and this requires the functions to be inline, and thus not virtual. Now that we're using (more) proper cpp, we can use virtual functions, and let someone else figure out which function we should be calling.

Tested this by playing mettagrid and mousing over things, and confirming that the observations still looked right.