Fix putting to correctly reduce inventory (#639)
We were decreasing the agent's inventory inside of an assert. The assert is fully stripped when not in tests (or similar); and this was also stripping the decrement.

Tested via the player.

I'm not sure if there's a leveraged way to write a test for this, given that the code worked as expected in a testing environment.