Switch events to use enum class (#647)
Refactored event handling system to use an enum class and a map for
better type safety and organization.

In another diff I ran into issues with enums and namespacing, and
learned that `enum class` exists seems to be generally better. This
incrementally switches us to use that, and does some other hygiene
around Events as well.

Tested with unit tests, and by running in `play` and seeing that mines
periodically create new ore.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>