Remove 1-hot encoding for type_ids (#708)
Replaced object type one-hot encodings with `agent`, `converter`, and `type_id` feature in observations.

This removes the 1-hot encoding of general type_ids, which mostly meant we were 1-hot encoding converter types. This makes our observation space more robust to changing converter types, and makes the CNN closer to what we're doing with tokens.

In basic smoke tests, this appeared to learn as fast as pre-ablation; though I haven't studied this thoroughly. The change is primarily on the principle of normalizing the data we're sending.