be consistent about using self._task (#920)
Removes a redundant parameter. Since `self._task` is already available
as a class member, there's no need to pass it as an argument to the
`_initialize_c_env` method. Moreover, passing it adds (or suggests) that
the parameters might be different.

There's more cleanup we could do, like the fact that `labels` are read
only off the first task. And it's not great that you need to remember to
set a new `_task` before calling init, and that it's done in 2 places.
But baby steps.

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>