Switch inventory to be maps (#1138)
Changes inventory representation from fixed-size vectors to maps. This
isn't strictly required to make things more configurable, but maps
better-match the mental model I think we should have for inventory (and
tokens), and should be more efficient if we have a vaguely-unbounded
number of types of things in inventory; or if we want inventory items to
not have to start at zero.

As per many PRs, this is an extraction from another PR.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210633750368373)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>