Add SequentialTrialTask (#1890)
Add SequentialTrialTask, and a demo config.

Introduces a new `SequentialTrialTask` class that allows agents to
progress through a sequence of trials, potentially based on how well
they're doing.

Note that "how we'll they're doing" should be a measure of "have they
in-context learned the thing we're trying to teach". I don't know if
these are the primitives we'll need, and it should work fine to leave
the advancement, regression, and max_trials all at default values.

I've smoke tested the code by training locally, but I / we don't have
the instrumentation we probably want to confirm this works as expected.
And I haven't written tests.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210953966325435)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>